Exibindo posts do firebase

+++ app/src/main/java/com/example/tsisocialapp/model/Post.kt
@@ -8,6 +8,6 @@ data class Post (
     var title: String,
     var text: String,
     var category: String,
-    var likes: Int,
+    var likes: Long,
     var image: String? = null
 )
